CurrentMatchコンポーネントとWaitingPlayersコンポーネントをuseSwapPlayerフックを使用してリファクタリングし、選手選択のロジックを整理
